+++ extensions/void/src/extension/extensionLib/readFileContentOfUri.ts
